requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Deprecated==1.2.14
Django==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
et_xmlfile==2.0.0
importlib_resources==6.4.5
inflection==0.5.1
iniconfig==2.0.0
kombu==5.4.2
numpy==2.1.3
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pluggy==1.5.0
prompt_toolkit==3.0.48
psycopg==3.2.3
PyJWT==2.9.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
wrapt==1.16.0
coverage>=7.4.0